Fetch API Implementation
The Implementation of getting data from the serverside and the Implementation of POST data to the server side which while uploading data using the POST converts the form data into JSON. While the fetch for getting data get JSON file whichyou can access.

We also created a new element while trying to show the output. We did this using JavaScript functions such as document.createElement() and document.append()'